hitl-new/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
asyncpg==0.30.0
pydantic==2.12.5
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.28.1
structlog==24.4.0
python-dotenv==1.0.1
python-multipart==0.0.20
websockets==13.1
pymupdf==1.24.0
